requirements.txt
streamlit
pand